Restructure TOCs for full Report.
After the titlepage, we give a concise overview of the full Report by
listing the main sections only.

Each start page of the two Volumes and the Appendices has its own local
TOC with the same structure as the original Report.

For dead-tree reading, we sprinkle in a few 'clear double pages' to
start on right sides of spreads.